LayoutTests/fast/dom/Range/getBoundingClientRect-on-collapsed-selection-range-expected.txt

hello
This tests calling getBoundingClientRect on a selection range when it is collapsed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect = getSelection().getRangeAt(0).getBoundingClientRect(); rect.x is 10
PASS rect.x is 10
PASS rect.y > 5 is true
PASS rect.width is 0
PASS rect.height > 15 is true
PASS successfullyParsed is true

TEST COMPLETE
